Now every graphics api implementation gets compiled and chosen at runtime
