+++ CMakeLists.txt
@@ -9,14 +9,5 @@ set(EXECUTABLE_OUTPUT_PATH "${CMAKE_BINARY_DIR}/bin")
 
 project(MagmaEngine VERSION 0.0.1 DESCRIPTION "The Magma Game Engine")
 
-if (APPLE)
-    set(MAGMA_RENDERER_VULKAN ON)
-    message(STATUS "Renderer: Vulkan for MacOS")
-    message(FATAL_ERROR "Vulkan renderer is currently not supported. You can you the Vulkan branch an experimental Vulcan renderer")
-else ()
-    set(MAGMA_RENDERER_OPENGL ON)
-    message(STATUS "Renderer: OpenGL 3")
-endif()
-
 add_subdirectory(Client)
 add_subdirectory(Magma)
