+++ Client/assets/Vulkan/Base.frag
@@ -1,0 +1,9 @@
+#version 450
+
+layout(location = 0) in vec4 fragColor;
+
+layout(location = 0) out vec4 outColor;
+
+void main() {
+    outColor = fragColor;
+}
