+++ Magma/src/Magma/Core/Application.cpp
@@ -4,29 +4,29 @@
 #include <Magma/Core/Input.h>
 #include <Magma/Renderer/Renderer.h>
 
+RenderingAPIType currentRenderingAPI = RenderingAPIType::OpenGL;
+
 namespace Magma
 {
     const std::vector<Vertex> vertices = {
-        {{ 0.5f,  0.5f, 0.0f}, {1.0f, 1.0f}},
-        {{ 0.5f, -0.5f, 0.0f}, {1.0f, 0.0f}},
+        {{0.5f, 0.5f, 0.0f}, {1.0f, 1.0f}},
+        {{0.5f, -0.5f, 0.0f}, {1.0f, 0.0f}},
         {{-0.5f, -0.5f, 0.0f}, {0.0f, 0.0f}},
-        {{-0.5f,  0.5f, 0.0f}, {0.0f, 1.0f}}
-    };
+        {{-0.5f, 0.5f, 0.0f}, {0.0f, 1.0f}}};
 
     const std::vector<uint32_t> indices = {
         0, 1, 3,
-        1, 2, 3
-    };
+        1, 2, 3};
 
     Mesh mesh;
     Material material;
-    glm::mat4 transform { 1.0f };
+    glm::mat4 transform{1.0f};
 
     void Application::Run()
     {
         mesh.Vertices = vertices;
         mesh.Indicies = indices;
-        material.Color = { 1.0f, 1.0f, 1.0f, 1.0f };
+        material.Color = {1.0f, 1.0f, 1.0f, 1.0f};
 
         m_Running = true;
         Logger::Init();
@@ -41,11 +41,11 @@ namespace Magma
         while (m_Running)
         {
             m_Window->Update();
-            
+
             Renderer::BeginFrame();
             Renderer::DrawMesh(mesh, *texture.get(), material, transform);
             Renderer::EndFrame();
-            
+
             m_Window->Draw();
         }
