+++ Magma/src/Magma/Core/Platform.h
@@ -37,3 +37,10 @@
 #else
 #error "Unknown platform!"
 #endif
+
+enum RenderingAPIType
+{
+    OpenGL
+};
+
+extern RenderingAPIType currentRenderingAPI; // Defined in Application.cpp
