+++ Magma/src/Magma/Platform/Linux/LinuxWindow.cpp
