+++ Magma/src/Magma/Platform/MacOS/MacOSWindow.cpp
